requirements.txt
deep_translator==1.6.0
discord.py==1.7.3
python-dotenv==0.19.2
requests==2.27.1
